code refactor - perf: add redirect before logging - refactor: change api folder structure - refactor: change linksController.js to two files - chore: add decoded token to body in tokenValidate middleware - fix: column name in linksController.js
