+++ models/linksController.js
@@ -1,54 +1,12 @@
 import {pool} from './database.js';
 
 /**
- * Fetch a row using token
- * @param {String} token - user's token
-*/
-export async function inspectToken(token){
-    let data = await pool.query('SELECT * FROM users WHERE token = $1', [token])
-
-    return data.rows[0]
-}
-
-
-/**
- * Updates a specific user's token
- * @param {String} email - user's email
- * @param {String} token - JWT token
-*/
-export async function updateToken(email, token){
-    await pool.query('UPDATE users SET token = $1 WHERE email = $2', [token, email])
-}
-
-/**
- * Inserts new user in the users table
- * @param {String} email - user's email
- * @param {String} password - user's password
- * @param {String} salt - password salt
-*/
-export async function createUser(email, password, salt){
-    await pool.query('INSERT INTO users(email, password, salt, created_at) VALUES($1, $2, $3, $4)', [email, password, salt, new Date(Date.now()).toDateString()])
-}
-
-
-/**
- * Fetch the data of a specific user's account using email
- * @param {String} email - email of the account
-*/
-export async function inspectUser(email){
-    const user = await pool.query('SELECT * FROM users WHERE email = $1', [email])
-
-    return user.rows[0]
-}
-
-
-/**
  * Creates new Link in the database
  * @param {String} userEmail - user's email
  * @param {String} shortened - the shortened version
  * @param {String} original - the original version of the url
 */
-export async function create(userEmail, shortened, original){
+export async function createLink(userEmail, shortened, original){
     const link = await pool.query('INSERT INTO links(owned_by, shortened, original, created_at) VALUES($1, $2, $3, $4)', [userEmail, shortened, original, new Date(Date.now()).toDateString()])
 
     if(link.rowCount == 0){
@@ -60,15 +18,15 @@ export async function create(userEmail, shortened, original){
  * Fetch original and id of a link using shortened code
  * @param {String} shortened - the shortened code
 */
-export async function inspect(shortened){
-    const link = await pool.query('SELECT id, original FROM links WHERE shortened = $1', [shortened])
+export async function inspectLink(shortened){
+    const link = await pool.query('SELECT original FROM links WHERE shortened = $1', [shortened])
 
     return link.rows[0]
 }
 
 
 /**
- * Inserts a visit to a specific link
+ * Inserts a visit to a specific link in the visits table
  * @param {String} shortened - the shortened code
  * @param {String} ip - visit ip
  * @param {String} country - visit country
@@ -79,7 +37,7 @@ export async function inspect(shortened){
 */
 export async function addVisit(shortened, ip, country, device, browser, os, visitedAt = new Date(Date.now()).toDateString()){
     const visit = await pool.query(
-        'INSERT INTO visits(link, ip, country, device, browser, os, clickedAt) '+
+        'INSERT INTO visits(link, ip, country, device, browser, os, visited_at) '+
         'VALUES($1, $2, $3, $4, $5, $6, $7)', 
         [shortened, ip, country, device, browser, os, visitedAt]
         )
