+++ routes/api/users/login.js
@@ -1,4 +1,4 @@
-import { inspectUser, updateToken } from '../../models/linksController.js';
+import { inspectUser, updateToken } from '../../../models/usersContrller.js';
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 
@@ -17,7 +17,7 @@ export default async function login(req, res){
             return res.status(400).json({error: 'email or password is incorrect'})
         }
 
-        let token = jwt.sign({email: user.email, password: user.password}, process.env.JWT_SECRET, {algorithm: 'HS256', expiresIn: '30s'})
+        let token = jwt.sign({email: user.email, password: user.password}, process.env.JWT_SECRET, {algorithm: 'HS256', expiresIn: '1d'})
         let refreshToken = jwt.sign({email: user.email, password: user.password}, process.env.JWT_REFRESH_SECRET, {algorithm: 'HS256'})
 
         await updateToken(user.email, refreshToken)
