+++ routes/api/users/token.js
@@ -1,4 +1,4 @@
-import { inspectToken } from '../../models/linksController.js';
+import { inspectToken } from '../../../models/usersContrller.js';
 import jwt from 'jsonwebtoken';
 
 /**
@@ -7,11 +7,11 @@ import jwt from 'jsonwebtoken';
 export default async function token(req, res){
     try{
         try{
-            var verified = jwt.verify(req.body.token, process.env.JWT_REFRESH_SECRET, {})
+            var verified = jwt.verify(req.body.token, process.env.JWT_REFRESH_SECRET)
         }catch(err){
             var verified = null
         }
-
+        
         if(!verified){
             return res.status(400).json({ error: 'token is invalid' })
         }
@@ -24,7 +24,7 @@ export default async function token(req, res){
 
         let newToken = jwt.sign({ email: verified.email, password: verified.password }, process.env.JWT_SECRET, {algorithm: 'HS256', expiresIn: '1d'})
 
-        res.setHeader('set-cookie', [`JWT_TOKEN=${newToken}; httponly`])
+        res.setHeader('set-cookie', [`JWT_TOKEN=${newToken}; max-age=86400; httponly; samesite=lax; path=/`])
         res.sendStatus(200)
     }catch(err){
         console.log(err)
