+++ routes/middlewares/tokenValidate.js
@@ -6,7 +6,8 @@ import jwt from 'jsonwebtoken';
 export default async function validate(req, res, next){
     try{
         try{
-            var isValid = jwt.verify(rqe.cookies.JWT_TOKEN, process.env.JWT_SECRET)
+            var isValid = jwt.verify(req.cookies.JWT_TOKEN, process.env.JWT_SECRET)
+            req.body.decoded = jwt.decode(req.cookies.JWT_TOKEN)
         }catch(err){
             var isValid = null
         }
